[Chore] 추억 카드 관련 코드 정리 및 파일 분리
