+++ Harmony/Harmony/Views/MemoryCard/MemoryCardDetailView.swift
@@ -160,41 +160,11 @@ struct MemoryCardDetailView: View {
             viewModel.loadMemoryCardDetail(id: memoryCardId)
             viewModel.getSummary(for: memoryCardId)
         }
-//        .onReceive(NotificationCenter.default.publisher(for: UIApplication.willEnterForegroundNotification)) { _ in
-//            viewModel.loadMemoryCardDetail(id: memoryCardId)
-//            viewModel.getSummary(for: memoryCardId)
-//        }
     }
 }
 
-struct LoadingView: View {
-    @State private var isAnimating = false
 
-    var body: some View {
-        VStack {
-            Image("moni-wholebody")
-                .resizable()
-                .scaledToFit()
-                .frame(width: 100, height: 100)
-//                .rotationEffect(Angle(degrees: isAnimating ? 360 : 0))
-//                .animation(Animation.linear(duration: 1).repeatForever(autoreverses: false), value: isAnimating)
-            
-                .scaleEffect(isAnimating ? 1.1 : 0.9)
-                .animation(Animation.easeInOut(duration: 0.8).repeatForever(autoreverses: true), value: isAnimating)
-            Text("모니가 상세 정보를 가져오고 있어요😀")
-                .font(.headline)
-                .padding(.top, 20)
-        }
-        .onAppear {
-            isAnimating = true
-        }
-    }
-}
 
 #Preview {
     MemoryCardDetailView(memoryCardId: 1, groupId: 1)
-}
-
-#Preview {
-    LoadingView()
 }
